requirements.txt
requests>=2.31
python-telegram-bot>=20,<22
nest_asyncio>=1.6
